requirements.txt
absl-py==0.15.0
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
Django==3.0
django-heroku==0.3.1
entrypoints==0.3
filelock==3.3.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.41.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
huggingface-hub==0.1.0
idna==3.2
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==4.1.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.19.0
psycopg==3.0b1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoingecko==2.2.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.2.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
selenium==4.0.0
Send2Trash==1.8.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.12.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.13.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webdriver-manager==3.5.2
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==6.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.0.0
